chore(runtime): migrate inline test scripts to vitest
Replaces the three hand-rolled `*.mjs` test scripts with vitest-based
`*.test.ts` files. Keeps testing lightweight — vitest as a single
devDep on `@flue/runtime`, default reporter, no plugins, no extra
fixtures.

# What changed

* **`vitest` ^4.1.6** added as devDep on `@flue/runtime`. Brings its
  own esbuild for TS transformation; no other tooling needed.
* **`vitest.config.ts`**: minimal — include `test/**/*.test.ts`,
  everything else default. Five lines of actual config.
* **`pnpm test`** + **`pnpm test:watch`** scripts on the package;
  root-level `pnpm test` shells out via turbo; `turbo test` task
  added with `dependsOn: ["^build"]` (test source-imports so the
  package itself does not need rebuilding, but cross-package
  imports — e.g. `cli/src/lib/cloudflare-wrangler-merge.ts` from
  the wrangler test — do).
* **Tests migrated to `*.test.ts`** with `describe`/`it`/`expect`:
  - `run-registry.test.ts` (was `.mjs` — InMemoryRunRegistry + bare
    /runs/:runId integration via Hono). 10 tests.
  - `cf-run-registry.test.ts` (was `.mjs` — SQL ops + REST router
    via `node:sqlite` shim). 15 tests.
  - `wrangler-merge-registry.test.ts` (was `.mjs` — pure-fn
    `computeFlueMigrations` / `mergeFlueAdditions`). 7 tests.

  Total: 32 tests across 3 files, ~290ms.

* **Old `.mjs` scripts deleted.** With them go the bespoke
  `test(name, fn)` harness, the manual `console.log('  ✓ ...')`
  output, the `biome-ignore-all lint/suspicious/noConsole`
  directives, and the `dist/`-import dance that was load-bearing
  because Node's strip-only TS loader couldn't parse parameter
  properties. Vitest's esbuild handles all of it; tests now read
  source directly.

# Why now

Pre-Phase 2 cleanup. The inline scripts were honest about their
"good-enough until we need real tooling" posture — that point is
here. Vitest is the smallest move that gets us:

* `describe`/`it` grouping (cf-run-registry was already a fake-test
  harness; this is the real one).
* Snapshot support (Phase 2 will need it for OpenAPI spec assertions).
* Watch mode for fast iteration.
* Built-in TS, no separate build step before testing.

Importantly, the migration is **scoped to `@flue/runtime`**. Other
packages (`@flue/cli`, `@flue/sdk`, etc.) gain no test infrastructure
yet — they don't have any tests and adding vitest there too would be
work without payoff. When they grow tests, install vitest locally
there too. No workspace-level vitest config, no project list, no
shared fixtures dir. Per the lightweight goal.

# Regression

* `pnpm --filter '@flue/runtime' test` → 32/32 pass, ~290ms.
* `pnpm --filter '@flue/runtime' check:types` clean.
* `pnpm --filter '@flue/cli' check:types` clean.
* `pnpm check:lint` 0 errors, 327 warnings (unchanged).
* CF example builds cleanly; generated `_entry.ts` still has the
  shared kebab→Pascal helper inlined.

# Notes on type fidelity in the migrated tests

The old `.mjs` tests passed `roles: []` and `defaultStore: undefined`
to `createFlueContext`, which the TS types reject (the contract
requires `Record<string, Role>` and a `SessionStore` respectively).
The `.ts` migration tightens these to `roles: {}` and
`new InMemorySessionStore()` — no behavior change (both happen to
exercise the same code path for an agent that doesn't use either),
and the test config now matches what real builds emit.

The `mergeFlueAdditions` return type is intentionally
`Record<string, unknown>` (it carries arbitrary wrangler config
shape); the wrangler test narrows via a one-line `as { ... }` cast
at the call site rather than poisoning the function's signature
with test-only types.
