+++ packages/runtime/package.json
@@ -44,6 +44,8 @@
 	"scripts": {
 		"build": "tsdown",
 		"check:types": "tsc --noEmit",
+		"test": "vitest run",
+		"test:watch": "vitest",
 		"prepublishOnly": "cp ../../README.md ."
 	},
 	"dependencies": {
@@ -61,6 +63,7 @@
 	"devDependencies": {
 		"@cloudflare/workers-types": "^4.20250410.0",
 		"tsdown": "^0.20.3",
-		"typescript": "^5.9.2"
+		"typescript": "^5.9.2",
+		"vitest": "^4.1.6"
 	}
 }
